Update 05-create-semantic-kernel-ai-assistant.md
The path to the Python folder is incorrect due to case sensitivity. The original path is named "Python" in the guide while the main repository is "python". This throws the "Set-Location: Cannot find path '/home/peter/mslearn-ai-semantic-kernel/Labfiles/05-ai-assistant/Python' because it does not exist." error when run. The correction changes the pathname from "Python" to "python".